add exception

+++ src/main/java/com/wang/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.wang.community.controller;
 
 import com.wang.community.dto.QuestionDTO;
-import com.wang.community.mapper.QuestionMapper;
 import com.wang.community.model.Question;
 import com.wang.community.model.User;
 import com.wang.community.service.QuestionService;
@@ -22,13 +21,13 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
-        model.addAttribute("title",question.getTitle());
-        model.addAttribute("description",question.getDescription());
-        model.addAttribute("tag",question.getTag());
-        model.addAttribute("id",question.getId());
+        model.addAttribute("title", question.getTitle());
+        model.addAttribute("description", question.getDescription());
+        model.addAttribute("tag", question.getTag());
+        model.addAttribute("id", question.getId());
         return "publish";
     }
 
@@ -38,33 +37,33 @@ public class PublishController {
     }
 
     @PostMapping("/publish")
-    public String doPublish(@RequestParam(value = "title",required = false) String title,
+    public String doPublish(@RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description", required = false) String description,
-                            @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id", required = false) Integer id,
+                            @RequestParam(value = "tag", required = false) String tag,
+                            @RequestParam(value = "id", required = false) Long id,
                             HttpServletRequest request,
                             Model model) {
 
-        model.addAttribute("title",title);
-        model.addAttribute("description",description);
-        model.addAttribute("tag",tag);
+        model.addAttribute("title", title);
+        model.addAttribute("description", description);
+        model.addAttribute("tag", tag);
 
         if (title == null || title == "") {
-            model.addAttribute("error","标题不能为空");
+            model.addAttribute("error", "标题不能为空");
             return "publish";
         }
         if (description == null || description == "") {
-            model.addAttribute("error","内容不能为空");
+            model.addAttribute("error", "内容不能为空");
             return "publish";
         }
         if (tag == null || tag == "") {
-            model.addAttribute("error","标签不能为空");
+            model.addAttribute("error", "标签不能为空");
             return "publish";
         }
 
         User user = (User) request.getSession().getAttribute("user");
         if (user == null) {
-            model.addAttribute("error","用户未登录");
+            model.addAttribute("error", "用户未登录");
             return "publish";
         }
